Add prefix option to findtoolchain script
Currently the findtoolchain script is only capable of finding a
directory that contains a potential toolchain for a target.
It was left to targetDefs files to defined the prefix.

As to make it possible to compile targets with other toolchains than
the default one, this updates findtoolchain to also take care of
determining the prefix.

Resolves: LE-5261
Change-Id: I8b7c1dacdc1f5a118c8525ee06e272d825868c15