[AR758x] Set legato image to 4k page size as Default
In the coming days, It will only use 4K page Nand flash device. As
require we should set the image to 4K page as default for UBI and
kernel image. For the old device with 2k page size, LK will auto
detect and switch the image from 4K to 2K. And if the Nand flash
device is 4K size, Lk will do nothing to the 4K image.

Notes: This function is only for UBI + squashfs.

Resolves: QTI9X28-243
Change-Id: I1dd4fcedbf1e61e11ef64a3301a9cef79aff50d8
Signed-off-by: Caupar Gu <CGu@sierrawireless.com>