Fix app MD5 being changed when build dir is different
Change the MD5 calculations to exclude parts of paths beyond the top
of the staging directory for both apps and systems.

Make the mk tools define the LE_FILENAME macro on the compiler
command line when compiling C or C++ code.

Change the logging macros to use STRINGIZE(LE_FILENAME) instead of
__FILE__.

Resolves: LE-5424, full
Change-Id: I4f9859bcf8685668821e330a1a2047d291cc507a