Positioning Fix On Demand API
Fix bug concerning the positioning « Fix On Demand » API.
If the positioning navigation handler is not subscribed thru
the le_pos_AddMovementHandler function, the position’s data
are always returned as out of range (LE_OUT_OF_RANGE),
even if the position is fixed.
Fix issue for verticalMagnitude equal to 0: the movement handler
is raised on acquisition update rate.

Resolves: LE-5465
Related: LE-5014
Change-Id: I134fbd3217d9b135e86d2b9ef715d9a081ae3f46