[avcService] Send registration update right after an asset is removed
We used to send registration update 15 seconds after a LWM2M operation.
This was done to aggregate multiple changes to asset data. However, When
an App is stopped and started quickly (started before reg.update) the state of
the device and the server could go out of sync. This situation can cause
observe from not working. This change sends a registration update
as soon an asset it removed.

Resolves: LE-4681
Change-Id: Ifeb90e6461fb1307367fa0a428387aae8b5b80f3