Replace usage of deprecated readdir_r
Since glibc 2.24, readdir_r is deprecated, which generates a warning that
is converted as an error since we have 'warning as error'.

According to the readdir/readdir_r documentation, applications should use readdir.
In modern implementations, including glibc, readdir is thread-safe, so there is
no reason to stick with readdir_r.

Resolves: LE-5482
Change-Id: I8b0ee9aca799754fabd366048ca763fa0ca13d94