Change return type of SafeRef & HashMap getters
SafeRef and HashMap both returned a "void const*" pointer from their
getter functions that take an iterator as input.  This is a const
pointer to a mutable value, *not* a mutable pointer to const value.
They return types have been updated so that the key getters return a
mutable pointer to a const value and the value getter return a mutable
pointer.

Resolves: LE-5646
Change-Id: I041b340d08e03c8cb38ba66971cf46dedcb7812e