[mangoh] Driver for mapping MangOH Green IoT slots
MDM9615 BSP supports the AR7 development platform. Although MangOH
boards are very similar to the AR7 development kit, the key
difference was in the addition of IoT slots, where additional IoT
cards could be plugged in. This makes the MangOH platform modular,
so additional code was required to detect card presence and load
drivers for daughter cards.

This driver implements the following functionality for MangOH
green:
 * Detection of IoT cards in slots
 * Parsing EEPROMs on cards
 * Adding devices listed in card EEPROMs

The functionality covers GPIO, I2C, SPI, SDIO, USB and UART cards.
Open issue remains in the provision of platform_data structs for
daughter cards, which is specific to detected devices and cannot
be handled in this (generic) module. For now, this issue has to
be worked-around by redoing daughter card drivers to include
platform data.

Special care was given to the possibility of upgrading EEPROM
format. EEPROM format version 1.0 is handled in this commit, but
future implementation needs to be backwards-compatible to older
EEPROM formats because daughter cards may have a mixture of
EEPROM versions.

Code attempts to follow kernel coding standard for device drivers
and does not follow Legato coding standard.

Resolves: LXSWI9X1517-51
Change-Id: I8f6a32a2eebea00bdb504af10fa6bc6b8768efe0
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>