[avcService] Allow user apps to request AVMS session
User apps can use the following APIs to request to open a lwm2m
session and to get notified.

 - le_avData_SessionRequest
 - le_avData_SessionRelease
 - le_avData_AddSessionStateHandler
 - le_avData_RemoveSessionStateHandler

A “custom arbitrator” can be rolled in to the control app using the following APIs.
If the session control handlers is not registered, a session can be opened anytime by user apps.

 - le_avc_AddSessionRequestEventHandler
 - le_avc_RemoveSessionRequestEventHandler

Resolves: LE-5423, LE-5509
Related: LE-4681
Change-Id: I3868999e8d178e3bd36ace1cd634c972d9733845