[powerMgr] Reference-count wakeup sources
Add a bit LE_PM_REF_COUNT in createOpts parameter of
lm_pm_NewWakeupSource() to manage a reference counted wakeup source.
Each time a lm_pmStayAwake() is called, the counter is incremented.
Each time a lm_pmrelax() is called, the counter is decremented and
the wakeup source is released only when counter reaches 0.

Resolves: LE-1837
Change-Id: I11e63d26aebc50f2e695c538240d7266f5a7fbda