[GNSS] Close NMEA pipe properly when interrupt signal occurs
When stopping a NMEA pipe flux with an interrupt signal (like CTRL-C),
a SIGPIPE signal is rised and the GNSS engine is restarted.

The man of the pipe mentions that if the read end of a pipe is closed,
then a write will cause a SIGPIPE signal to be generated for the calling
process.

The SIGPIPE signal is now blocked to the posDaemon. Any write to the pipe
will only cause a write error.

Resolves: LE9XXX-192
Change-Id: I47f15025b8c311bea37b24e321d78cbc37d4a698