[GNSS] Add a note to le_gnss_InjectSuplCertificate API
A note is added to le_gnss_InjectSuplCertificate API explaining the
following case :
In case the SUPL certificate size given as input parameter is bigger than
the Maximum SUPL certificate size, it is a fatal error and the function
will not return.

Additionally, integration tests verifying SUPL legato functions are added.

Resolves: LE9XXX-290, LE9XXX-288
Change-Id: I45962317695f2bac9f99391cabb93ef5224f2148