Provide error when Legato sources are incomplete
When the https://github.com/legatoproject/legato-af repository is cloned
by itself, or archives generated by GitHub are downloaded, the project
cannot build because it is missing some important directories such
as platformAdaptor.
These directories are actually repositories that are either in the
official release package
(such as https://downloads.sierrawireless.com/legato/1607/legato-16.07.0.tar.bz2 )
or provided by repo.

This patch updates the Makefile to check for these directories, and in
case they are not present it will provide the following error:
 'platformAdaptor' directory is missing, which means these Legato sources have not been downloaded properly.
 Please refer to https://github.com/legatoproject/legato-af#clone-from-github

Resolves: LE-5897
Change-Id: I796ddfd51ca4583bc23ea6316ead0de389eda53b