Move and fix some C framework UTs
Looks like these tests have been left unbuilt and unexecuted for a
little while.

The memPool UTs fails with error: Error finding pools by name: 530
Therefore I created ticket LE-5909 to track the fix of that test.
In the meantime the failing code has been commented out.

The log UTs are also failing, as it seems the API has been updated
while the tests were not. Fixing of this test is in LE-5910.

Resolves: LE-5908
Related: LE-5909 LE-5910
Change-Id: I36da7ae6eb4a158ca2d3c8918a10eb9a7e5fbf58