[drivers] Enable driver build within Legato
This relatively large commit adds the ability to build kernel
driver modules within Legato. For this purpose, keyword "sources:"
was added to .mdef files to specify all source files that comprise
the kernel module. A user can now choose to use the "sources:" or
"preBuilt:" section, but using both would generate an error.

If a "sources:" keyword is specified, then Legato generates a
driver Makefile and builds the driver as part of Legato build.

The "cflags:" and "ldflags:" sections list all options passed to
the compiler and linker during driver build, respectively.

A handy feature is the "LEGATO_KERNELROOT" environment variable.
If this environment variable is defined within the "buildVars:"
section of .sdef file, it points the build to a custom kernel
location. If - as part of default behaviour - this variable is
not specified, the build uses the kernel directory from Legato
tool installation.

Change-Id: I5c8a5675cbef6a9d0141e632ce4b422eb14a68f1
Resolves: LE-6007
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>