Add generation of wp76xx CWE images to Yocto build
Issue Description:
Currently, Legato CWE has PID of 9X28. In ticket QTI9X07-55, we
changed the logic of 9X07 to only accept PID with Y921. We must
change the PID of Legato CWE to Y921 for WP76xx so that Legato
can be used.

Root cause:
The default PID for WP76xx is 9X28.

Solution:
Change the default PID for WP76xx to Y921

Testing:
Compiled image_src and check the header of the image. The header of
the image shows Y921.

Change-Id: I6c233f90740df60c408c4f5c1c9b14bf1dfc01f5
Related: QTI9X07-55
Resolves: QTI9X07-203