Provide secStoreGlobal API + unit tests
This creates a /global namespace in the secure storage, accessible
by a service called secStoreGlobal, which reuses the le_secStore.api
already used for the storage of per-app or per-user data.

In order to connect to the service, applications should declare
themselves as client of the le_secStore.api, and bind themselves to
the secStore.secStoreGlobal interface.

I also created some unit tests to reuse the existing target testing
on localhost.

Resolves: LE-6067
Change-Id: Id5f028104c6f1d010ca95a7586062a8906d5617f