Provide build parameter LEGATO_APP_AVC
The idea is to be able to select the AVC service that should
be used for the build between 3 flavours:
- lwm2mcore: the new service that doesn't rely on the modem or a PA
  (aka AVC 2.0)
- legacy: the deprecated service that relies on QMI PA AVC
  (aka AVC 1.0)
- none: does not build the system with an AVC service

On mdm9x28/mdm9x40, lwm2mcore is the default as these platforms do not
have an embedded AVC service.

As the APIs for the new lwm2mcore agent are going to be updated in the
near future, the existing ones (le_avc.api & le_avdata.api) are moved to
a new sub-directory within interfaces/airVantage/ called 'legacy'.
People that wish to continue using the legacy agent would have to modify
their app to point to interfaces/airVantage/legacy/ instead of
interfaces/airVantage/.

The new repository for the lwm2mcore solution is using an older version
of the APIs so these files are taking a step back in time.

This patch also introduce a way to declare 'external dependencies',
which is volontarily vague, and which allows the notification of make
targets that will be generated before the generation of the system.

Resolves: LE-6080
Related: LE-6079
Change-Id: I37bfbda37b93ea33b96aba053bc2b85e324b34b3