Add ability to include .Xdef files.
Add a new processing directive '#include' which includes the contents of
another file at the location of the directive. Unlike the C '#include'
it does not need to be on its own line.  An '#include' directive was used
over an 'include:' section for maximum flexibility in composing .Xdef
files.

Throw an error if a buildVar which have been used in an '#include'
directive is redefined later to a different value.  This avoids confusing
cases where an
environment variable has two different values in different
parts of a .Xdef file.  buildVar sections cannot simply be processed first
(as we are doing to avoid the same issue in other sections) since an
included file could have its own buildVar sections.

Split defFile_t into two classes:
 - defFile_t which represents a top-level .Xdef file
 - defFileFragment_t which represents any file, including included files
This was done to separate things only present in top level files
(like top-level sections) from things which are part of any file
(like the token list).

Refactor the lexer so Lexer_t::Pull will return the next important
tokens. Whitespace and comments are still added to the
defFileFragment_t they appear in, but are not returned by Lexer_t::Pull.
This simplifies the parser as there is now a single location for
expanding processing directives, and ignoring whitespace and comments.
This changes the format somewhat as there are a few locations where
whitespaces and comments are now allowed where they were not previously,
but this is acceptable since any file recognized by the old parser will
still be recognized by the new parser.

Change-Id: I0d4d4cfcc52eb0f25fbd277b4d48c093785a722c
Resolves: LE-4250