Add conditional directive processing to Xdefs
Adds conditional directive (#if, #elif, #else, and #endif) processing to Xdef
files.  As with includes, this is done in the lexer for maximum flexibility
as to where these conditionals can be included.

The initial phase only supports a single condition per directive
(no &&, || or similar) and supports a limited set of conditions:
 - two tokens are equal
 - a file exists, or
 - a directory exists.

These are sufficient for the needs of LE-4970.  More conditions may be added
later as needed.

Resolves: LE-4517
Change-Id: I27d524d7588904e66e1c6fede6aff7c766fbd552