Update le_updateCtrl API.
le_updateCtrl_LockProbation() may fail to take a lock if the probation period
is not active.  The function should return a status in that case to indicate
the lock was not taken.  For example, in this case
le_updateCtrl_UnlockProbation() should not be called.

Add a new API le_updateCtrl_GetSystemState() to get the current state of the
system (good, bad or probation).  This can only be called by a client
if a probation lock is held, or if updates are deferred as otherwise the
system state could change at any time (including between when the function
is called and its return value is checked).

Resolves: LE-6183
Change-Id: I233025a9c3402a153b331c652f9649318a0be1a0