Use virt platform for target testing
The idea is to provide a way to package a virt runtime as a Docker
image.
Then it is easy to instanciate a virtual target, run the tests against it
and tear it down without having to manage a real target.

To build the Docker image, you would need to do:
$ make virt
$ cd targetFiles/virt
$ ./build.sh

To run the tests:
DEST_IP=172.17.0.3 TARGET_TYPE=virt ./run-tests target-ut

Resolves: LE-5908 LE-4559 LE-6531 LE-4135
Change-Id: Ib330d10847412a575f48e24216725c089c8cd693