[cm tool] Update 'cm data connect' connection detection
Updating connection detection from google.com to akamai.com
as to make sure that it works from China, and using HTTP instead of
ICMP (ping) as this can be limited on some networks.

The 'Checking if device if up ...' is currently not doing anything
as the $iface variable is initialized later in the function.

Resolves: LE-6528
Change-Id: I202d8dfaa0f201b81e0efaa30e3c402fce7fee90