Disconnect audio stream before parameters release
At the end of an audio playback, the destruction of the audio stream
involved releasing the audio parameters and then disconnecting the
stream.
This sequence must be inverted since the audio parameters are
needed to disconnect the stream porperly.

This patch inverts the call sequence in order to disconnect the audio
stream first and then release the parameters.

Resolves: LE-6412
Related: LE9XXX-768
Change-Id: Ie5a81d28a7cf84256be2ced1d9def1cf4d0727d3