Fix occasional crash on Legato shutdown.
If an app crashes when the system is shutting down, do not attempt to restart
the app.

If the supervisor receives a spurious shutdown notification, ignore it instead
of asserting.  Such notifications can be generated due to races between an
app being shutdown and the app crashing or quitting.

Add an API to determine if the Legato framework is shutting down.

Resolves: LE-6505
Change-Id: Ib984d1166a5d63603ce580b7137b9d3b54e8ea5a