Do not crash on NULL input pointers.
If a NULL pointer was passed in an array input, the generated code would
crash if the array size was non-zero.  Check and reject this case.

Mark input string pointers as nonnull so we can catch incorrect useage of
these APIs at compile time.  As this causes the compiler to assume
these parameters are nonnull, remove the checks (clang issues a tautology
warning on the check).

Array pointers cannot be marked as nonnull
since a NULL pointer is valid for 0-size arrays.

Rework a few cases where NULL pointers were being incorrectly passed in string
parameters.  This removes a few tests which now generate compile errors --
which is good as it shows some incorrect code which previously had to
be detected at runtime is now found at compile time.

Resolves: LE-6799
Change-Id: I988c857b03d6e20a099fab1a8443d1c203963566