Handle 0-sized buffer provided by client.
We are now passing number of characters accepted to server, instead of buffer
size, to be consistent with all other size usage in IPC protocol.  However
some tests will pass 0 as buffer size, which means not even the terminating
NULL is allowed.  Use UINT32_MAX as buffer size to represent this case.

I'm somewhat unclear if this is even a valid test case.  The API format
specifies that a shorter out string can be used; however in this case
there isn't enough space to store any valid string, as all C strings
are at least 1 character long (the terminating NULL).

Resolves: LE-6802
Change-Id: If3e3c1586d65cd2fab88beee2e28b44780f4e66d