[AT Services] Add atBinder app
Add atBinder application, it's a server app that can receive AT cmds on a
socket from any platform-dependent process.
Path of the socket is "/tmp/atBinder".
This app also opens a device on this socket with the le_atServer_Open()
function to start the AT commands monitoring with the atServer service.

Resolves: LE-6607
Change-Id: Idea657311af37f778095cbfd760c3461ffcc44c1