Change error on binding unreferenced interface to a warning
Previous versions of Legato would allow binding an interface not referenced
by any compoment.  This was the documented way of binding interfaces in
externally built components (e.g. legacy apps being ported to Legato).

With the introduction of binary apps, legacy apps should use the binary
app mechanism of including the interface in the extern: requires: section.
For this release the old mechanism should still be allowed but deprecated.

Resolves: LE-6890
Change-Id: I67b83fe9f8edea46f4a1a3fcdb476404137f8d0b