Allow playing back audio files in loop in ifgen code
Due to a change in the C code generator, audio clients which pass
LE_AUDIO_NO_FD parameter to rewind a file in order to play it back in
loop are killed during playback.
This patch makes the LE_AUDIO_NO_FD parameter valid again in the
generated C code in order to allow playing back audio files in loop.

Resolves: LE-6899
Change-Id: Iaa51dac800be3e9e81a30ca606ef4650cf4da0c7