Fix 'config export' if the destination file does not exist
For instance doing:
  config export tree:/ newfile
doesn't work, while doing:
  config export tree:/ existingfile
works.

This is because the tool tries to make the file path absolute before
giving it to the le_cfgAdmin API.
However at the moment it does not handle the case where the file has
not been created yet.

I provided some target tests but the ACL tests are failing because
logs from sandboxed apps are not present after '/etc/init.d/syslog restart'.

Resolves: LE-6877
Related: LE-6986
Change-Id: I8c33bbdf1db85ff4d5e305d92234d64b3228a1db