Re-add support for old-style handler parameters.
The old ifgen supported parameters with type 'handler'.  This was removed from
the new ifgen as it's more consistent to treat handlers as a user-defined type
and use the handler name as the parameter type.

However using 'handler' as a parameter type was never deprecated, so cannot
be removed yet.  Re-add support, but generate a deprecation warning and
mark as deprected in the documentation.

Resolves: LE-6829
Change-Id: Ic113bd1e418692ed56f381a0582294b9271a61e6