Update run-tests to run target UTs on virtual platform
This provides a way to run target UTs against a virtual target that
is instanciated and brought down just for the purpose of running these
tests.

'simu start' cannot be used at the moment to start the target as it
would result in a non-default SSH port, while at the moment Legato tools
only support the default SSH port (= 22).

Therefore docker is used instead as it also allows to manage the image
provisioning and bring up/bring down quite easily.

Also update the findtoolchain script to handle 'virt' as a target.

Resolves: LE-5908
Change-Id: I0c027e08c9333cdc04717222e89fd4ae969776f7