Prevent infinite resets of Legato (version 2)
Track Legato boot loops, and if a boot loop is detected:

 - Restore from golden master, if another system is the current system;
 - Halt if the golden master is current.

This attempts to recover from some undetected bad upgrades, or at least
preserve flash memory if recovery is not possible.

A read-only system has nowhere to store a boot count, so do not do
boot loop prevention on a read-only system.

Resolves: LE-4587
Change-Id: I7864c8c623ed1ac9270115a8edf0b9a6b83bc0b4