Enforce unique process names.
The supervisor & config tree already assume process names are unique; however
mkTools does not.  As a result, mkTools will allow multiple processes with
the same name; however these are silently dropped (and only the last one is
used) when the app is installed on the target.

app runProc did allow creating multiple processes with the same name.  To
be consistent with predefined apps, restrict to unique names; adding a suffix
to the end of the executable name to make the name unique.

Resolves: LE-5759
Change-Id: Ibe6be4389d8e6a178c1825dd85e335f6041d547b