Fix detection of changes for proprietary apps in releaselegato
The .app files are tarballs and the current implementation to detect
changes relies on doing a hash of the content.
However since we cannot control the order in which files are packaged
in the tarball, this currently doesn't work properly and it thinks that
the content changes everytime.

This replaces the detection by an actual extraction of the tarball and a
comparaison of the directories.

Resolves: LE-6518
Change-Id: I92e25ab1fb4c913c3092edc95bed76ee20111c5e