Limit unpacked string to actual string buffer size.
The user could pass in a string buffer which is smaller than the expected
string buffer size.  In such a case, ensure the string never overflows
the actual buffer passed in.

The server should ensure they never pass a larger string than the client
expects, but this check ensures a malicious client cannot cause a buffer
overflow.

Resolves: LE-7340
Change-Id: I56a6f337002eaea8bb007342e6d60369456f5c9e