Check bundled file existance after build
With the introduction of the externalBuild: section, there's a
mechanism to generate new files which need to be bundled during a
build.  So files bundled in a bundle: file: section should only be
checked to see if they exist when finally bundling the file, not when
the ninja build script is generated.

Note this does not apply to bundle: dir: sections since dependencies
can only be generated correctly if all bundled files are known in
advance.

Resolves: LE-7440
Change-Id: I93464517e2ad828e65ba00bb0dae3aa842b97ab7