Add info printing to testFwRand to prevent timeout
The random generation depends on the machine speed, entropy available,
current load and the time necessary to collect samples for this test
therefore change widely.
The ctest timeout is set to 120s of inactivity which is sometime not enough to
complete the test.
This change adds a message printed every few samples:
[47.5%] Collecting ... 19000000 samples [655.2 ksamples/s]
[50.0%] Collecting ... 20000000 samples [645.2 ksamples/s]

It shows the progression, amount of samples collected and  the average
collection speed.

Resolves: LE-7454
Change-Id: I198b463b0cc50d5766454657109a28340599eca5