Only package gdbserver in devMode if available
The build currently fails if gdbserver is not available in the
toolchain, which is quite common for any toolchain that does not
descend from the SWI Yocto distribution.
Instead of failing, this skips the bundling and produce a runtime
warning.

Since the only reason devMode was not made available on localhost
is because gdbserver was not available at the right place, this is
also removing the build switch.
As it turns out, the localhost build can be a bit picky on some
hosts so I reworked a bit the tcf-agent start-up sequence.

Resolves: LE-7484
Change-Id: I3dc545121bf9e6da9da8a3bbb5e983ce4322b513