Update toolchain determination
Toolchain determination was not very flexible because it
could only be controlled by X_TOOLCHAIN_DIR and X_TOOLCHAIN_PREFIX
environment variables, with the added kludge of looking at the
USE_CLANG env var when building for localhost.

Also, toolchain determination was pretty much completely
undocumented.

The toolchain determination has been moved to the start-up process
of all of the mk tools, and the details are now stored in the
BuildParams_t object. It can also be controlled using CC,
CXX, LD, AR, AS, STRIP, OBJCOPY, READELF variables, or preferably
by X_CC, X_CXX, X_LD, etc (where X is the target ID, like WP85 or
wp85).  LEGATO_SYSROOT and X_SYSROOT can also be used.

The X_TOOLCHAIN_DIR and X_TOOLCHAIN_PREFIX variables are also
still supported and will be used if one of the other variables
is not set. x_TOOLCHAIN_DIR and x_TOOLCHAIN_PREFIX support
has also been added.

The documentation has been updated (written) for all this.
A few other minor doc fixes were also made in related pages.

The splitdebug script run by the mk tools was modified to
use the STRIP, READELF and OBJCOPY env vars instead of
TOOLCHAIN_DIR and TOOLCHAIN_PREFIX and the targetDefs
make script was modified to export STRIP, READELF and OBJCOPY
for use by mkexe/splitdebug when building the framework.

It was also noticed that error checking was missing from
part of splitdebug.  This was corrected.

Resolves: LE-7574
Change-Id: I6f0b3852d446e0b0c6306f8206a9a09b67977e6d