Switch to clang _Nonnull for nullability in ifgen
Previously ifgen would indicate a parameter could not be NULL through
the gcc attribute __attribute__((nonnull)).  However this had the
undesirable side-effect of also causing the optimizer to assume a
parameter was nonnull, leading to any attempts at defensive programming
with this parameter being thrown out by the optimizer.

Now use clang's _Nonnull keyword instead which will issue a warning if
a NULL parameter is passed, but does not optimize out NULL pointer checks
on the parameter.  This means GCC will no longer check for null pointers
in these parameters; however there does not appear to be a method in GCC
of getting the desired check without the undesirable optimization changes.

Resolves: LXSWIREF-91
Change-Id: Ied63eec58133b815a77893158c5fe668e7785719