Define generator data structure.
Instead of plain-old sequential logic for generating system/app/etc.
define a list of generators which will be applied to generate a build
for a particular target (i.e. Linux/RTOS/etc).

This is a preparatory step for increased flexibility in the future.
Going forward this list could be updated dynamically by plugins;
defined in a target-specific data file; etc.

Resolves: LE-7556
Change-Id: I201869b0168858c8d790d14e234b0f7a71532c8c