Fix build fail with clang from outside /usr/bin
The targetDefs makefile was setting the TOOLCHAIN_DIR
to the directory in which clang resided whenever
USE_CLANG=1. But then READELF, STRIP, and OBJCOPY are
set to $(TOOLCHAIN_DIR)$(TOOLCHAIN_PREFIX)readelf, etc.
So, if clang was installed somewhere other than /usr/bin/,
the build would fail when it tried to run readelf.

Resolves: LE-7627
Change-Id: Ia4adcdf908dfde6a8c8747f98a4039cecb147c66