Close audio streams in a specific order
During an audio use-case teardown, the frontend file play and capture
streams are closed in a non-deterministic order after the audio path
has been reset. On some platforms, this behaviour induces a bug where
the DSP is not able to release audio resources as expected.
This patch inverts the teardown sequence in order to close frontend
file play and capture streams first.

Resolves: LE-7499
Change-Id: I6a282117e3f3fb876e41a3bbeabc89df3ecbd66f