[Docs] Fix up format/spelling in  docs
This fix is the combination of a couple of small tickets
plus an audit of building the documentation of things
that were broken refs and section headings.  These are
all very minor changes mostly spelling and formatting
changes.
This will get done periodically from now on to improve
the readability of the docs.

Resolves: LE-7621 LE-7614 LE-7071 LE-7780 LE-7752
Change-Id: I5a42a5a417dc55ab8251a6a44b1846fe3b7918b5