Fix OOB write in le_fs_Read
Revert some changes introduced recently in le_fs_Read to add
a null character at the end of the buffer.

le_fs is not to process strings but binaries in general, so the output
is not necessarily null terminated.
The current code causes a buffer overflow as 'rc' can hold the size of the
array if read() fills the buffer.

Resolves: LE-7494
Change-Id: I08b8ffe3234404deab9382969001787ba18e5a54