Prevent audio xrun when all samples have been read
When there are no more samples to be played during a PlaySamples
procedure, the playback thread can fall into an audio underrun if the
ALSA software stack was excpecting additionnal frames; this happens
when the shortage of frames is due to latency during playback.
This patch prevent these audio underrun cases by writing silence
frames to the ALSA driver when all samples have been read.

Resolves: LE-7843
Change-Id: I61867340c151b565b65d35f431a0238d76fbeb72