Fix app control issues
There was some documentation errata in the le_appCtrl.api that
resulted in an incorrect error message being printed when an
app's container could not be created in response to 'app start X'
on the command line.

While looking into this, it was discovered that there were a lot
of things for which documentation was missing completely in the
same .api file.

Error messages are also improved so that grepping the logs
for the name of the app that failed to start will find more
helpful troubleshooting information that it did previously.

Some minor code quality improvements were also made along the
way to reduce confusion for future maintainers and to check
some error codes that were being ignored.

Resolves: LE-8207 LE-6783
Change-Id: I71b0decfa97cdd6781ae7ead5a81d43e54fc7c51