Fix CMake non-existent dependencies
When non-existent dependencies exists, CMake policy CMP0046 kicks in
and fails the build.

Dependencies in log tests are from the pre-ninja era and not needed anymore
as the framework is build before the build of tests starts.

Build targets in fwupdate are conditioned by LEGATO_TARGET and therefore
do not exist for every conditions.

Resolves: LE-8213
Change-Id: I6a22af07b7e70070d6e92c09a9b5d342674090c0