Change default values of CLIR and CUG for le_mcc
Add new values to the CLIR/CUG enums to be able to not include these
fields in the CC Setup message.
And set these values to be the default ones for le_mcc_Start().

When CLIR field is not set, API le_mcc_GetCallerIdRestrict() returns
LE_UNAVAILABLE.

Resolves: LE-8486
Related: LEAR7M-107
Change-Id: I22f543fd87db287caec16fa9790e8e48cec55b16