Update le_pack_PackString for klocwork issue
Just like strncpy, this function doesn't know the size of the
source string, which can potentially result in a read overflow
on the input buffer.

Static code analyzer can report this issue, which is deliberatly
ignored since calls to le_pack_PackString need to be NULL-terminated.

Also adding some unit tests to check the behavior of le_pack.

Resolves: LE-8567
Change-Id: I24c9180571742b21ca486ac5ed2671580633d626