Fix detection of failures in run-tests
The return code was not properly propagated, resulting
in a script that exists with 0 despite some failed target tests.

I reworked the logic a little bit and aligned the behavior of
host and target testing.
Now it prints 'Host tests failed' and 'Target tests failed' in
case of failure.

Tested both cases by introducing some errors in the tests and the
exit code at the end of the tests is different from 0.

Resolves: LE-8641
Change-Id: I2cf4e75f6c62dfae55d760ffab8d0df0aa0e1bcf