Allow test application to play audio samples in loop
The audioPlaybackRec test application did not allow playback in loop
when using the "PB_SAMPLES" test-case because the underlying
le_audio_PlaySamples() API does not have rewind capability (this
capability is offered with le_audio_PlayFile() through "PB"
test-case).
For the sake of providing an example of how to play audio samples in
loop despite the le_audio_PlaySamples() API capabilities, this patch
allows the audioPlaybackRec test application to run the "PB_SAMPLES"
test-case in loop when "LOOP" is specified on the command line
interface.

Resolves: LE-8625
Change-Id: I3a7405529c696c9e8406dde6ee9dc93e17f03efa