Return audio service unavailability error code at API level
When audio service requests fails, the client application can not know
whether the failure is caused by an initialization failure or any
other kind of failure. Being able to discriminate between the two
would allow the client application to try recovering by restarting the
service in case of audio initialization failures.
This patch allows the audio service API to return the LE_UNAVAILABLE
error code when the failure is due to an initialization issue.

Resolves: LE-8547
Change-Id: Iea687c00530b5f91fb740ce5c8915f0b19f26b83