Update apt-get in travis file
There are multiple list of packages to install with apt-get in
the repo, that got out-of-sync:
- 'zsh' was still mentioned in Doxygen
- travis list was totally outdated

Also removed vim from the list as it shouldn't really be up
to a framework to give a preference on the editor that should
be used to edit it (and everyone uses vim anyway :p).

I tried to update the travis build to also build and run tests
but this is going to take a bit more time, as some of the test
case explictly uses gcc to build the code.
For the test part, it seems that the travis perf can be quite
limited on some aspects, resulting in flaky tests (the tests
themselves are likely flaky, not sure that travis is to blame
here).

Resolves: LE-8768
Change-Id: I5150a4975c86936f2a2445aca60f6a1644bce084