Fix coverity warning related with CHECKED_RETURN
The warning is due to ignored return value from read system
call on /proc/<id>/cmdline file.
It is safe to assume that read on /proc/PROC-ID/cmdfile will
not returns 0 or -1 specially when above file is open successfully.
Fixed the CHECKED_RETURN error by typecasting the return value of read to void.

Resolves: LE-8772
Change-Id: Id95ebb1009962e7de07f7dee281dc8981a18f9e7