Add timer to check the exit status for stopping an APP
When going to stop an APP, "MarkAppAsStopped" will try to
check if all the process exit or not for that APP, and without
doing any dealy between every try. It may easily reach to the
max try, and then trigger reboot. This patch add timer for the
check to give time for APP to exit.

Resolves: LE-8750
Relates: LEAR7M-144
Change-Id: I7cd77532ca11d191034d55be4b2284d6bce9847c