GPRS Data connection is not consistent.
Description:
Customer had observed a behaviour change in the cm data service
on WP8548 (Release 14 legato 16.10.3) when compared to Release 9
legato 16.01.

Root cause:
Data session is terminated if the network does not provide DNS
addresses. While integrating wifi into dcsDaemon this behaviour
change was introduced.

Solution:
If the network provides a valid gateway address but not a DNS
address we should not terminate the session.
This change starts a timer if the DNS values were not obtained
and reattempts to acquire them after timer expiry.

Resolves: QTI9X1555-1238
Change-Id: I884c8a89b78f6ff102618a6d17029b9636a05c70
Signed-off-by: lkmthiy <lkmthiy@sierrawireless.com>