Fix buffer overflow in le_rand_GetBuffer
I noticed that the quantity asked from getrandom/read (buffer size)
is not updated.
So when these functions returns less quantity than what we asked for,
we are actually asking for the same quantity on the next run while
we should really be asking for less since the pointer on the buffer
has moved.

Resolves: LE-9306
Change-Id: I3bbd4bbb512b75c0609fc9d644b9741a43553f11