[virt] Workaround qemu crash
It seems like qemu crashes when performing reboots, especially
in cloud environment such as Azure or GCE.

This works around it by restarting qemu on errors.
It restarts it only if the process has not exited 'immediately'
(ie, that it took more than 5 secs for the process to exit with
an exit code != 0). This is to avoid infinite loops on configuration
errors.

It is possible to disable that restart through RESTART_ON_ERROR=0.

This allows to mitigate the issue has no solution has been found
so far.

Resolves: LE-9516
Related: LXSWIREF-401
Change-Id: I46397acdaf3fc004006e6cfc312229fc622360b9