DCS retry WiFi after Cellular Connection Fails
When a cellular connection is made and after a while it is lost (Remove the SIM
for instance).  In this case the connection service is disconnected and there
will be no attempt to used another technology. The user have to send a command
to connect to server again.  The DCS has been updated to attempt to reconnect
again using the next technology (WiFi).

Resolves: LE-9305
Change-Id: Ic65b4252ba16667d63704d4213f699fd5f008ee0