[UT] Reinforce SIGFPE test
This test appears to be quite weak at the moment
and can fail to generate a SIGFPE, which appears to be
what we want.

Loop over the triggers and print the result as to
hopefully succeed in generating a FPE.

Same goes for ABRT that might not be generated on
the first double-free.

Resolves: LE-9583
Related: FARM-1488
Change-Id: I74bfc01832e99ca6255cf3f292a9e71681d1977f