Do not execute testFwRand as a perf test on localhost
When executing the test for le_rand APIs, the UT collect
samples and eventually determine if they are random enough.

It appears that some of the VMs that we are running this on
do not produce numbers that are random enough, so this fails
from time to time, and quite a lot these days.

Instead this transform the test to execute on localhost without
testing for performance.

The test is added to the list of on-target testing, where it does
test for performance.

Resolves: LE-9582
Change-Id: I41cea6e17be528ce6fbe6dcb2514a6031f358028