Timer API enhancements
To make it easier to use repeating timers that need their interval
to change from time-to-time, make it possible to set the interval
while it is running.

Provide Get functions for interval and time remaining
until next expiry.

Add a second thread to the timer unit tests.

Make tests exit with non-zero exit status when they fail.

Resolves: LE-9648 LE-9649 LE-9650
Change-Id: I59b45c02542716bf021ca34651b102bf25bb1971