General impovements to the def file handling.
Adding the ability for .sdef files to specifiy seperate search
within a sdef.  Specifically, the search paths for
applications/components/modules/source code are now seperate.

The command line parameter --source-search will update all of these
paths in unison to maintain backwards compatability and to simplify
that interface.

Made the CURDIR more context aware, it will not reflect the def file
properly, even if it is a fragment included from another location.

Resolves: LE-4026 LE-8693 LE-8699
Change-Id: I6a435b49e35cfd44c640212482ee7021a72ce4eb