Update few 3rd party components
The idea is to make some applications independent from
the content of the rootfs by packaging the libraries they
depend on with the application.

For instance, an app like the AVC depends on zlib, curl, openssl ...
If the rootfs goes back to a previous version, or an updated version,
the AVC daemon might not work properly.
By making the app independent, it becomes more robust but takes slightly
more space.

Resolves: LE-7217 LE-9643 #32
Change-Id: I09950e1cff3c62d2e9048cca936b92ffd5becc1f